make some changes

+++ src/varasto.py
@@ -1,7 +1,7 @@
 class Varasto:
-    def __init__(self, tilavuus, alku_saldo = 0):
+    def __init__(self, tilavuus, alku_saldo=0):
         self.tilavuus = (tilavuus + abs(tilavuus))/2
-        self.saldo = (alku_saldo + abs(alku_saldo))/2 if alku_saldo <= tilavuus else tilavuus
+        self.saldo = float((alku_saldo + abs(alku_saldo))/2) if alku_saldo <= tilavuus else tilavuus
 
     # huom: ominaisuus voidaan myös laskea. Ei tarvita erillistä kenttää viela_tilaa tms.
     def paljonko_mahtuu(self):
@@ -19,7 +19,7 @@ class Varasto:
         if maara < 0:
             return 0.0
         if maara > self.saldo:
-            ota_kaikki=self.saldo
+            ota_kaikki = self.saldo
             self.saldo = 0.0
             return ota_kaikki
